+++ examples/discovery-search-app/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.1-beta.19](https://github.com/watson-developer-cloud/discovery-components/compare/discovery-search-app@1.0.1-beta.18...discovery-search-app@1.0.1-beta.19) (2020-02-12)
+
+**Note:** Version bump only for package discovery-search-app
+
+
+
+
+
+## [1.0.1-beta.18](https://github.com/watson-developer-cloud/discovery-components/compare/discovery-search-app@1.0.1-beta.17...discovery-search-app@1.0.1-beta.18) (2020-02-12)
+
+**Note:** Version bump only for package discovery-search-app
+
+
+
+
+
+## [1.0.1-beta.17](https://github.com/watson-developer-cloud/discovery-components/compare/discovery-search-app@1.0.1-beta.16...discovery-search-app@1.0.1-beta.17) (2020-02-12)
+
+**Note:** Version bump only for package discovery-search-app
+
+
+
+
+
+## [1.0.1-beta.16](https://github.com/watson-developer-cloud/discovery-components/compare/discovery-search-app@1.0.1-beta.15...discovery-search-app@1.0.1-beta.16) (2020-02-11)
+
+**Note:** Version bump only for package discovery-search-app
+
+
+
+
+
+## [1.0.1-beta.15](https://github.com/watson-developer-cloud/discovery-components/compare/discovery-search-app@1.0.1-beta.14...discovery-search-app@1.0.1-beta.15) (2020-02-11)
+
+**Note:** Version bump only for package discovery-search-app
+
+
+
+
+
+## [1.0.1-beta.14](https://github.com/watson-developer-cloud/discovery-components/compare/discovery-search-app@1.0.1-beta.13...discovery-search-app@1.0.1-beta.14) (2020-02-10)
+
+**Note:** Version bump only for package discovery-search-app
+
+
+
+
+
+## [1.0.1-beta.13](https://github.com/watson-developer-cloud/discovery-components/compare/discovery-search-app@1.0.1-beta.12...discovery-search-app@1.0.1-beta.13) (2020-02-08)
+
+**Note:** Version bump only for package discovery-search-app
+
+
+
+
+
 ## [1.0.1-beta.12](https://github.com/watson-developer-cloud/discovery-components/compare/discovery-search-app@1.0.1-beta.11...discovery-search-app@1.0.1-beta.12) (2020-02-06)
 
 **Note:** Version bump only for package discovery-search-app